refactor(controllers): abstract out reusable code
